requirements.txt
awesome-slugify==1.6.5
Django==1.10.2
django-ckeditor==5.1.1
Pillow==3.4.2
regex==2016.10.22
Unidecode==0.4.19
